requirements.txt
beautifulsoup4==4.11.1
python-dotenv==0.21.0
selenium==4.5.0
tinydb==4.7.0
undetected_chromedriver==3.1.6
